Onboarding flow sequence with credential flow
