+++ components/org.wso2.carbon.identity.scim2.common/src/main/java/org/wso2/carbon/identity/scim2/common/utils/SCIMCommonUtils.java
@@ -24,7 +24,6 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.CarbonConstants;
 import org.wso2.carbon.context.PrivilegedCarbonContext;
-import org.wso2.carbon.identity.application.authentication.framework.util.FrameworkUtils;
 import org.wso2.carbon.identity.claim.metadata.mgt.ClaimMetadataHandler;
 import org.wso2.carbon.identity.claim.metadata.mgt.ClaimMetadataManagementService;
 import org.wso2.carbon.identity.claim.metadata.mgt.exception.ClaimMetadataException;
@@ -32,7 +31,6 @@ import org.wso2.carbon.identity.claim.metadata.mgt.model.ExternalClaim;
 import org.wso2.carbon.identity.claim.metadata.mgt.model.LocalClaim;
 import org.wso2.carbon.identity.core.ServiceURLBuilder;
 import org.wso2.carbon.identity.core.URLBuilderException;
-import org.wso2.carbon.identity.core.context.IdentityContext;
 import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
 import org.wso2.carbon.identity.core.util.IdentityUtil;
 import org.wso2.carbon.identity.handler.event.account.lock.constants.AccountConstants;
@@ -62,17 +60,11 @@ import org.wso2.charon3.core.schema.AttributeSchema;
 import org.wso2.charon3.core.schema.SCIMConstants;
 
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 /**
  * This class is to be used as a Util class for SCIM common things.
@@ -1161,118 +1153,4 @@ public class SCIMCommonUtils {
         }
         return Boolean.parseBoolean(considerServerWideUserEndpointMaxLimitProperty);
     }
-
-    public static boolean isExecutableUserProfileUpdate(Map<String, String> userClaimsToBeModified,
-                                                        Map<String, String> userClaimsToBeDeleted) throws UserStoreException {
-
-        if (!MapUtils.isEmpty(userClaimsToBeModified) || !MapUtils.isEmpty(userClaimsToBeDeleted)) {
-
-            return hasAnyNonFlowInitiatorClaims(userClaimsToBeDeleted.keySet()) ||
-                    hasAnyNonFlowInitiatorClaims(userClaimsToBeModified.keySet());
-        }
-
-        return false;
-    }
-
-    private static boolean hasAnyNonFlowInitiatorClaims(Set<String> claimUriList) throws UserStoreException {
-
-        for (String claimUri : claimUriList) {
-
-            if (!isFlowInitiatorClaim(claimUri)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public static boolean isFlowInitiatorClaim(String claimUri) throws UserStoreException {
-
-        ClaimMetadataManagementService claimMetadataManagementService = SCIMCommonComponentHolder
-                .getClaimManagementService();
-        String tenantDomain = IdentityContext.getThreadLocalIdentityContext().getTenantDomain();
-
-        try {
-            Optional<LocalClaim> localClaim = claimMetadataManagementService.getLocalClaim(claimUri, tenantDomain);
-            return localClaim.isPresent() && localClaim.get().getFlowInitiator();
-        } catch (ClaimMetadataException e) {
-            throw new UserStoreException(String.format("Error while reading claim meta data of %s", claimUri), e);
-        }
-    }
-
-    public static boolean isExecutableUserProfileUpdate(Map<String, String> userClaimsExcludingMultiValuedClaimsToBeModified,
-                                                        Map<String, String> userClaimsExcludingMultiValuedClaimsToBeDeleted,
-                                                        Map<String, List<String>> simpleMultiValuedClaimsToBeAdded,
-                                                        Map<String, List<String>> simpleMultiValuedClaimsToBeRemoved)
-            throws UserStoreException {
-
-        if (!MapUtils.isEmpty(userClaimsExcludingMultiValuedClaimsToBeModified) ||
-                !MapUtils.isEmpty(userClaimsExcludingMultiValuedClaimsToBeDeleted) ||
-                !MapUtils.isEmpty(simpleMultiValuedClaimsToBeAdded) ||
-                !MapUtils.isEmpty(simpleMultiValuedClaimsToBeRemoved)) {
-
-            return hasAnyNonFlowInitiatorClaims(userClaimsExcludingMultiValuedClaimsToBeModified.keySet()) ||
-                    hasAnyNonFlowInitiatorClaims(userClaimsExcludingMultiValuedClaimsToBeDeleted.keySet()) ||
-                    hasAnyNonFlowInitiatorClaims(simpleMultiValuedClaimsToBeAdded.keySet()) ||
-                    hasAnyNonFlowInitiatorClaims(simpleMultiValuedClaimsToBeRemoved.keySet());
-        }
-
-        return false;
-    }
-
-    /**
-     * Populate the multi-valued claims to be modified by adding the values to be added and
-     * removing the values to be removed from the existing claim values of a particular multi-valued claim of a user.
-     *
-     * @param existingClaimsOfUser      Existing claims of the user.
-     * @param multiValuedClaimsToAdd    Multi-valued claims to be added.
-     * @param multiValuedClaimsToDelete Multi-valued claims to be deleted.
-     * @return Multi-valued claims to be modified.
-     */
-    public static Map<String, String> getSimpleMultiValuedClaimsToModify(
-            Map<String, String> existingClaimsOfUser,
-            Map<String, List<String>> multiValuedClaimsToAdd,
-            Map<String, List<String>> multiValuedClaimsToDelete) {
-
-        String multiAttributeSeparator = FrameworkUtils.getMultiAttributeSeparator();
-        Map<String, String> multiValuedClaimsToModify = new HashMap<>();
-
-        Map<String, List<String>> existingClaimsAsList = existingClaimsOfUser.entrySet().stream()
-                .collect(Collectors.toMap(
-                        Map.Entry::getKey,
-                        entry -> new ArrayList<>(
-                                Arrays.asList(entry.getValue().split(Pattern.quote(multiAttributeSeparator))))));
-
-        Function<String, List<String>> getMutableClaimList = claimURI -> {
-            if (multiValuedClaimsToModify.containsKey(claimURI)) {
-                return new ArrayList<>(Arrays.asList(
-                        multiValuedClaimsToModify.get(claimURI).split(Pattern.quote(multiAttributeSeparator))));
-            }
-            return new ArrayList<>(existingClaimsAsList.getOrDefault(claimURI, new ArrayList<>()));
-        };
-
-        // Add the values to be added to the existing values of each multi-valued claim and
-        // construct the new claim value joining the values
-        if (multiValuedClaimsToAdd != null) {
-            multiValuedClaimsToAdd.forEach((claimURI, valuesToAdd) -> {
-                List<String> currentValues = getMutableClaimList.apply(claimURI);
-                currentValues.addAll(valuesToAdd);
-                multiValuedClaimsToModify.put(claimURI, String.join(multiAttributeSeparator, currentValues));
-            });
-        }
-
-        // Remove the values to be removed from the existing values of each multi-valued claims
-        // and construct the new claim value joining the values
-        if (multiValuedClaimsToDelete != null) {
-            multiValuedClaimsToDelete.forEach((claimURI, valuesToDelete) -> {
-                List<String> currentValues = getMutableClaimList.apply(claimURI);
-                if (!currentValues.isEmpty()) {
-                    currentValues.removeAll(valuesToDelete);
-                    multiValuedClaimsToModify.put(claimURI, String.join(multiAttributeSeparator, currentValues));
-                }
-            });
-        }
-
-        return multiValuedClaimsToModify;
-    }
-
 }
